refactored video delegate
