--- sample.config.ini
@@ -1,16 +1,0 @@
-# Do not store plain text credentials in production environments
-
-# Zoom Meeting SDK Client ID
-client-id=""
-
-# Zoom Meeting SDK Client Secret
-client-secret=""
-
-# Use a join-url or a meeting-id and password
-join-url=""
-
-# Choose your bots display name
-display-name="Sample Bot"
-
-[RawAudio]
-file="meeting-audio.pcm"
