+++ sample.config.toml
@@ -1,0 +1,17 @@
+#### !! Do not store plain text credentials in production environments !!
+#### File Format: https://cliutils.github.io/CLI11/book/chapters/config.html#configure-file-format
+
+# Zoom Meeting SDK Client ID
+client-id=""
+
+# Zoom Meeting SDK Client Secret
+client-secret=""
+
+# Use a join-url or a meeting-id and password
+join-url=""
+
+[RawAudio] 
+file="meeting-audio.pcm"
+
+#[RawVideo]
+#file="meeting-video.yuv"
